tests/legacy/build_constraints: Split Go asm test into own package
The native "go build" tool compiles assembly in a Cgo package with
the C compiler, and compiles assembly in a pure Go package with the
Go assembler. The build_constraints test previously mixed Go
assembler files with Cgo files. This caused the native go test to
fail with:

package using cgo has Go assembly file asm_linux_amd64.s

Splitting this into two separate packages fixes this.

This problem was reported in:
https://github.com/bazelbuild/rules_go/issues/2006

This will fix failing tests in my attempt to fix that issue:
https://github.com/bazelbuild/rules_go/pull/3648
