+++ tests/legacy/build_constraints/BUILD.bazel
@@ -16,10 +16,6 @@ go_library(
         # Check that tags are observed.
         "tag_l.go",
         "tag_unknown.go",
-        # Check that constraints apply to assembly files.
-        "asm_arm64.s",
-        "asm_linux_amd64.s",
-        "asm_unknown.s",
         # Check that constraints apply to cgo files.
         "cgo_linux.go",
         "cgo_unknown.go",
