+++ tests/legacy/build_constraints_go_asm/BUILD.bazel
@@ -1,0 +1,20 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_test(
+    name = "go_default_test",
+    size = "small",
+    srcs = ["build_constraints_go_asm_test.go"],
+    embed = [":go_default_library"],
+)
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        # Check that constraints apply to assembly files.
+        "asm_arm64.s",
+        "asm_linux_amd64.s",
+        "asm_unknown.s",
+    ],
+    cgo = False,
+    importpath = "github.com/bazelbuild/rules_go/tests/build_constraints_go_asm",
+)
