+++ tests/legacy/build_constraints_go_asm/asm_linux_amd64.s
