+++ tests/legacy/build_constraints_go_asm/asm_unknown.s
